server/requirements.txt
python-dotenv
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-deepgram
livekit-plugins-cartesia
livekit-plugins-elevenlabs